% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featheriseR_csv.R
\name{featheriseR_csv}
\alias{featheriseR_csv}
\title{A featheriseR Function}
\usage{
featheriseR_csv(ip)
}
\arguments{
\item{input}{file, csv with headers.}
}
\description{
This function allows you to convert files to feather format.
}
\examples{
featheriseR_function()
}
\keyword{feather}

