% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featheriseR_xls.R
\name{featheriseR_xls}
\alias{featheriseR_xls}
\title{A featheriseR Function}
\usage{
featheriseR_xls(ip, sheet.num)
}
\arguments{
\item{input}{file, xls with headers.}

\item{input}{sheet number.}
}
\description{
This function allows you to convert files to feather format.
}
\examples{
featheriseR_function()
}
\keyword{feather,}
\keyword{xls}

