% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featheriseR_xlsx.R
\name{featheriseR_xlsx}
\alias{featheriseR_xlsx}
\title{A featheriseR Function}
\usage{
featheriseR_xlsx(ip, sheet.name)
}
\arguments{
\item{ip}{input file, xlsx with headers.}

\item{sheet.name}{input sheet name.}
}
\description{
This function allows you to convert files to feather format.
}
\keyword{feather,}
\keyword{xlsx}

