% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featheriseR_csv.R
\name{featheriseR_csv}
\alias{featheriseR_csv}
\title{A featheriseR Function}
\usage{
featheriseR_csv(ip)
}
\arguments{
\item{ip}{input file}
}
\value{
Outputs a feather file to the same path as input
}
\description{
This function allows you to convert a csv file to feather format.
}
\examples{
featheriseR_csv(system.file("examples", "applicants1.csv",
package = "featheriseR"))
}
\keyword{csv}
\keyword{feather,}

