% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featheriseR_xls.R
\name{featheriseR_xls}
\alias{featheriseR_xls}
\title{A featheriseR Function}
\usage{
featheriseR_xls(ip, sheet.name)
}
\arguments{
\item{ip}{input file, xls with headers.}

\item{sheet.name}{input sheet name.}
}
\value{
Outputs a feather file to the same path as input
}
\description{
This function allows you to convert an xls file to feather format.
}
\examples{
featheriseR_csv(system.file("examples", "applicants.xls", package = "featheriseR"))
}
\keyword{feather,}
\keyword{xls}

