% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featheriseR_xls.R
\name{featheriseR_xls}
\alias{featheriseR_xls}
\title{A featheriseR Function}
\usage{
featheriseR_xls(ip, sheet.name)
}
\arguments{
\item{ip}{input file, xls with headers.}

\item{sheet.name}{input sheet name.}
}
\description{
This function allows you to convert files to feather format.
}
\keyword{feather,}
\keyword{xls}

