% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featheriseR_xlsx.R
\name{featheriseR_xlsx}
\alias{featheriseR_xlsx}
\title{A featheriseR Function}
\usage{
featheriseR_xlsx(ip, sheet.name)
}
\arguments{
\item{ip}{input file, xlsx with headers.}

\item{sheet.name}{input sheet name.}
}
\value{
Outputs a feather file to the same path as input
}
\description{
This function allows you to convert an xlsx file to feather format.
}
\examples{
featheriseR_xlsx(system.file("examples", "applicants.xlsx",
package = "featheriseR"), sheetName = 'applicants')
}
\keyword{feather}
\keyword{xlsx}

